FIx in case of drawing after loading npy file
Maybe I am wrong, or this can conflict with other parts of the code, but when loading the matrix of cells from a npy file, __cover_walls function will draw on x,y axis. But I noticed that x and y variables in input to the function are computed counting respectively the number of rows and the number of cols and maybe they should be swapped when drawing.